+++ src/app/component/visit-summary/diagnosis/diagnosis.component.ts
@@ -10,10 +10,7 @@ import {
   animate,
   keyframes,
 } from "@angular/animations";
-import { MatSnackBar } from "@angular/material/snack-bar";
-declare var getEncounterProviderUUID: any,
-  getFromStorage: any,
-  getEncounterUUID: any;
+declare var getEncounterUUID: any;
 
 @Component({
   selector: "app-diagnosis",
@@ -62,9 +59,8 @@ export class DiagnosisComponent implements OnInit {
   constructor(
     private service: EncounterService,
     private diagnosisService: DiagnosisService,
-    private snackbar: MatSnackBar,
     private route: ActivatedRoute
-  ) {}
+  ) { }
 
   ngOnInit() {
     this.visitUuid = this.route.snapshot.paramMap.get("visit_id");
@@ -91,9 +87,7 @@ export class DiagnosisComponent implements OnInit {
   onSubmit() {
     const date = new Date();
     const value = this.diagnosisForm.value;
-    const providerDetails = getFromStorage("provider");
-    const providerUuid = providerDetails.uuid;
-    if (providerDetails && providerUuid === getEncounterProviderUUID()) {
+    if (this.diagnosisService.isSameDoctor()) {
       this.encounterUuid = getEncounterUUID();
       const json = {
         concept: this.conceptDiagnosis,
@@ -107,20 +101,18 @@ export class DiagnosisComponent implements OnInit {
         this.diagnosisList = [];
         this.diagnosis.push({ uuid: resp.uuid, value: json.value });
       });
-    } else {
-      this.snackbar.open("Another doctor is viewing this case", null, {
-        duration: 4000,
-      });
     }
   }
 
   delete(i) {
-    const uuid = this.diagnosis[i].uuid;
-    this.diagnosisService.deleteObs(uuid).subscribe((res) => {
-      this.diagnosis.splice(i, 1);
-      if(this.diagnosis.length === 0) {
-        this.isDataPresent.emit(false);
-      }
-    });
+    if (this.diagnosisService.isSameDoctor()) {
+      const uuid = this.diagnosis[i].uuid;
+      this.diagnosisService.deleteObs(uuid).subscribe((res) => {
+        this.diagnosis.splice(i, 1);
+        if (this.diagnosis.length === 0) {
+          this.isDataPresent.emit(false);
+        }
+      });
+    }
   }
 }
